Reintroduced the graphic terminal as a user task.
Overall Changes
---------------
- The old graphics package found under "HAVK_Kernel.Graphics.*" has
  been recovered and modified into a ring 3 operating system program.
  It currently functions as a simple text displayer by printing
  whatever comes to port 404, but a fully working command-line is now
  closer than ever before.
- The PS/2 driver has been made into its own task. As of now, it
  simply sends individual key press data to any task called "Terminal"
  and to its port specified prior.
- A few helper subprograms have been created in the base operating
  system package so less time is spent on managing system call states.

+++ source/operating_system/framebuffer_tester/main.c
@@ -127,7 +127,7 @@ uint64_t main(void)
 	framebuffer = (uint32_t *)arguments.argument_1;
 	framebuffer_size = arguments.argument_2;
 	framebuffer_width = arguments.argument_3 >> 32;
-	framebuffer_height = arguments.argument_3 & UINT16_MAX;
+	framebuffer_height = arguments.argument_3 & UINT32_MAX;
 	scanline_width = arguments.argument_4;
 	pixel_width = scanline_width / framebuffer_width;
 	pixel_type = arguments.argument_5;
